resolved string empty in register
