+++ AcercaDeNosotros.html
@@ -35,7 +35,7 @@
         </div>
       </section>
     </main>
-    <script src="js/contacto.js" type="module"></script>
+    <script src="js/login.js"></script>
     <script src="js/bootstrap.min.js"></script>
     <script
       src="https://kit.fontawesome.com/7d2e342abf.js"
