+++ Error404.html
@@ -33,7 +33,7 @@
         <a class="btn404" href="/index.html" role="button">Go to home</a>
       </div>
     </section>
-    <script src="js/contacto.js" type="module"></script>
+    <script src="js/login.js"></script>
     <script src="js/bootstrap.min.js"></script>
     <script
       src="https://kit.fontawesome.com/7d2e342abf.js"
