+++ Register.html
@@ -433,7 +433,7 @@
       </section>
     </main>
     <script src="js/bootstrap.min.js"></script>
-    <script src="js/login.js" type="module"></script>
+    <script src="js/login.js" ></script>
     <script
       src="https://kit.fontawesome.com/7d2e342abf.js"
       crossorigin="anonymous"
