+++ index.html
@@ -313,7 +313,7 @@
         </div>
       </section>
     </main>
-    <script src="js/login.js" type="module"></script>
+    <script src="js/login.js" ></script>
     <script src="js/bootstrap.min.js"></script>
     <script
       src="https://kit.fontawesome.com/7d2e342abf.js"
