+++ login.html
@@ -223,7 +223,7 @@
     <main>
       <section class="container-fluid ContentRegisterItems">
         <div class="containerBtnLogin bg-dark my-5 px-5">
-          <h1 class="text-light text-center m-4">Crea Tu Usuario</h1>
+          <h1 class="text-light text-center m-4">Registrate</h1>
           <form id="FormLogin">
             <div class="mb-1">
               <label for="Nombre" class="form-label text-light">Nombre</label>
